implement the set config and refactor
